chore: Remove unused files and update Lead Connector integration in Ava API
