+++ app/api/ava.py
@@ -2,7 +2,6 @@ from fastapi import APIRouter, Depends
 from app.security import get_api_key
 from fastapi import APIRouter
 
-
 router = APIRouter()
 
 @router.get("/chat")
