+++ app/api/webhook.py
@@ -1,19 +1,17 @@
 from datetime import datetime
 import json
-from typing import List
+from typing import List, Optional
 from fastapi import APIRouter, Request
 from loguru import logger
 from pydantic import BaseModel, Field
 
-from app.integrations.leadconnector import (
-    LeadConnector,
-)
-from app.services.message_service import LeadConnectorMessageingService
+from app.services.lead_connector_messaging_service import LeadConnectorMessageingService
+
 
 class LeadConnectorWHTypeInboundMessage(BaseModel):
     type: str = Field(..., example="InboundMessage")
     locationId: str = Field(..., example="l1C08ntBrFjLS0elLIYU")
-    attachments: List[str] = Field(..., example=[])
+    attachments: Optional[List[str]] = Field(..., default_factory=list(), example=[])
     body: str = Field(..., example="This is a test message")
     contactId: str = Field(..., example="cI08i1Bls3iTB9bKgFJh")
     contentType: str = Field(..., example="text/plain")
@@ -23,7 +21,9 @@ class LeadConnectorWHTypeInboundMessage(BaseModel):
     messageType: str = Field(..., example="SMS")
     status: str = Field(..., example="delivered")
 
+
 router = APIRouter()
+
 
 def is_lc_location_accepted(location_id: str) -> bool:
     """
@@ -36,7 +36,7 @@ def is_lc_location_accepted(location_id: str) -> bool:
         bool: True if the location is accepted, False otherwise.
     """
     file_path = ".config/accepted_locations.json"
-    
+
     try:
         with open(file_path, "r") as file:
             file_data = json.load(file)
@@ -44,7 +44,7 @@ def is_lc_location_accepted(location_id: str) -> bool:
     except FileNotFoundError as exc:
         logger.error("File not found")
         raise FileNotFoundError("File not found") from exc
-        
+
     if location_id in accepted_locations:
         return True
     else:
@@ -54,9 +54,12 @@ def is_lc_location_accepted(location_id: str) -> bool:
 @router.post("/leadconnector")
 async def leadconnector(request: Request):
     request = await request.json()
+    logger.info(f"webhooked by leadconnector location id: {request['locationId']}")
 
     if not is_lc_location_accepted(request["locationId"]):
-        logger.warning(f"Location {request['locationId']} not accepted, skipping this WebHook event")
+        logger.warning(
+            f"Location {request['locationId']} not accepted, skipping this WebHook event"
+        )
         return
 
     request_type = request["type"]
@@ -70,15 +73,14 @@ async def leadconnector(request: Request):
 
         wh_message = LeadConnectorWHTypeInboundMessage(**request)
         lc_messaging_service = LeadConnectorMessageingService()
-        
+
         # if the incomming message is a special code, process it and return, dont go further
         if lc_messaging_service.process_special_codes(
-            message=wh_message.body,
-            conversation_id=wh_message.conversationId
+            message=wh_message.body, conversation_id=wh_message.conversationId
         ):
             return
-        
+
         # if the message is not a special code, respond to the message
-        lc_messaging_service.respond(
-            contact_id= wh_message.contactId,
-            conversation_id= wh_message.conversationId)
+        lc_messaging_service.respond_to_inbound_message(
+            contact_id=wh_message.contactId, conversation_id=wh_message.conversationId
+        )
